LayoutTests/inspector/styles/inactive-properties.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    InspectorTest.selectNodeAndWaitForStylesWithComputed("inspected", step1);

    function step1()
    {
        InspectorTest.dumpSelectedElementStyles(false, false);
        InspectorTest.completeTest();
    }
}

</script>
<style>
#inspected {
    text-align: left;
    text-align: bar;
    text-align: right;
}
</style>
</head>

<body onload="runTest()">
<p>
Tests that effectively inactive properties are displayed correctly in the sidebar.
</p>

<div id="container">
    <div id="inspected" align="left">Test</div>
</div>

</body>
</html>
