LayoutTests/inspector/styles/inactive-properties-expected.txt

Tests that effectively inactive properties are displayed correctly in the sidebar.

Test
[expanded] 
display: block;
    div - block user agent stylesheet
text-align: right;
    /-- overloaded --/ #inspected - left inactive-properties.html:20
    /-- overloaded --/ #inspected - bar inactive-properties.html:20
    #inspected - right inactive-properties.html:20
    /-- overloaded --/ div[Attributes Style] - -webkit-left 

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
#inspected  { (inactive-properties.html:20)
/-- overloaded --/ text-align: left;
/-- overloaded --/ text-align: bar;
text-align: right;

[expanded] 
div[Attributes Style]  { ()
/-- overloaded --/ text-align: -webkit-left;

[expanded] 
div  { (user agent stylesheet)
display: block;
